rename: companies data file name
